Back to functioning OAuth1 flow
